Update educationclass-list-members.md
When a class contains more than 100 members, Microsoft Graph returns a @odata.nextLink property in the response that contains a URL to the next page of results. If that property is present, continue making additional requests with the @odata.nextLink URL in each response, until all the results are returned, as described in paging Microsoft Graph data in your app.